# Use MOI.add_constrained_variable when creating variables