Replacement for #3862

See also https://github.com/jump-dev/HiGHS.jl/pull/245