@odow 2024-11-01
## Before

```Julia
julia> using JuMP, HiGHS

julia> function main()
           model = direct_model(HiGHS.Optimizer())
           set_silent(model)
           set_string_names_on_creation(model, false)
           @variable(model, 1 <= x[i in 1:10_000] <= 2)
           return model
       end
main (generic function with 1 method)

julia> @time main();
  0.025593 seconds (30.06 k allocations: 2.307 MiB)

julia> using BenchmarkTools

julia> @benchmark main()
BenchmarkTools.Trial: 198 samples with 1 evaluation.
 Range (min … max):  24.018 ms …  31.237 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     25.140 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   25.339 ms ± 981.654 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

       ▁▂▂▄ ▁▂█▇▆▄                                              
  ▃▁▁▆▆████▅████████▃▄▁▃▆▁▃▁▃▆▁▁▁▃▃▃▃▃▃▁▁▃▁▁▁▁▁▃▁▃▁▁▃▁▁▁▁▁▁▃▁▃ ▃
  24 ms           Histogram: frequency by time         29.2 ms <

 Memory estimate: 2.31 MiB, allocs estimate: 30056.

(hgh) pkg> st
Status `/private/tmp/hgh/Project.toml`
  [87dc4568] HiGHS v1.10.2
  [4076af6c] JuMP v1.23.3
```

## After

```Julia
julia> using JuMP, HiGHS

julia> function main()
           model = direct_model(HiGHS.Optimizer())
           set_silent(model)
           set_string_names_on_creation(model, false)
           @variable(model, 1 <= x[i in 1:10_000] <= 2)
           return model
       end
main (generic function with 1 method)

julia> @time main();
  0.016782 seconds (30.06 k allocations: 2.307 MiB)

julia> using BenchmarkTools

julia> @benchmark main()
BenchmarkTools.Trial: 284 samples with 1 evaluation.
 Range (min … max):  16.344 ms … 29.127 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     17.092 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   17.609 ms ±  1.672 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄█▆▆▆▆▄▄▃▁▂ ▁                                                
  ██████████████▆█▆▆▆▁▄▆▆▆▆▄▄▁▄▄▄▁▁▁▁▆▄▁▄▄▁▁▄▁▁▁▁▁▄▁▄▁▁▁▁▁▁▁▄ ▆
  16.3 ms      Histogram: log(frequency) by time      25.5 ms <

 Memory estimate: 2.31 MiB, allocs estimate: 30056.

(hgh2) pkg> st
Status `/private/tmp/hgh2/Project.toml`
  [87dc4568] HiGHS v1.10.2 `https://github.com/jump-dev/HiGHS.jl.git#master`
  [4076af6c] JuMP v1.23.3 `https://github.com/jump-dev/JuMP.jl.git#od/add-constrained-variable`
```

## Change

So from 24ms to 16ms, which is about 1/3 faster, or exactly what we predicted might be the case on the call.

## Single-sided bounds

Single-side bounds are twice as fast. From 16ms to 8ms:

```julia
julia> using JuMP, HiGHS

julia> function main()
           model = direct_model(HiGHS.Optimizer())
           set_silent(model)
           set_string_names_on_creation(model, false)
           @variable(model, x[i in 1:10_000] >= 0)
           return model
       end
main (generic function with 1 method)

julia> @time main();
  0.049627 seconds (41.81 k allocations: 3.119 MiB, 62.90% compilation time)

julia> using BenchmarkTools

julia> @benchmark main()
BenchmarkTools.Trial: 274 samples with 1 evaluation.
 Range (min … max):  16.755 ms …  22.301 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     18.170 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   18.237 ms ± 583.379 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

                      ▂  ▄▁█▁▃▄▆▄▃▂ ▄▁▃   ▁▂                    
  ▄▁▁▁▁▁▁▁▃▄▁▄▃▅▅▆▇█▆▅█▇▇██████████▇████▅▅██▆▆▇▃▃▅▃▃▄▃▄▄▄▁▄▁▁▄ ▄
  16.8 ms         Histogram: frequency by time         19.6 ms <

 Memory estimate: 2.31 MiB, allocs estimate: 30056.
```
to
```Julia
julia> using JuMP, HiGHS

julia> function main()
           model = direct_model(HiGHS.Optimizer())
           set_silent(model)
           set_string_names_on_creation(model, false)
           @variable(model, x[i in 1:10_000] >= 0)
           return model
       end
main (generic function with 1 method)

julia> @time main();
  0.041924 seconds (41.62 k allocations: 3.094 MiB, 73.63% compilation time)

julia> using BenchmarkTools

julia> @benchmark main()
BenchmarkTools.Trial: 468 samples with 1 evaluation.
 Range (min … max):   8.908 ms … 375.817 ms  ┊ GC (min … max): 0.00% … 93.33%
 Time  (median):      9.863 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   10.672 ms ±  16.926 ms  ┊ GC (mean ± σ):  7.02% ±  4.31%

     ▃▁█▃▅▆  ▁  ▂ ▁▃ ▄ ▆▁▇▃ ▂▄▁▃▃▁                              
  ▃▃▆██████▅██▇▇█▇████▇████▅██████▅▄▆▆▃▇▆▄▄▃▃▃▄▃▄▁▃▁▃▃▃▃▃▁▃▃▁▃ ▄
  8.91 ms         Histogram: frequency by time         11.7 ms <

 Memory estimate: 2.31 MiB, allocs estimate: 30056.
```