@odow 2024-11-01
So this is an initial backwards compatible minimal change that is a good performance improvement, but only partially solves the problem.

@joaquimg, you could always write your variables as `x in MOI.Interval(l, u)` to get another speedup from
```julia
julia> function main()
           model = direct_model(HiGHS.Optimizer())
           set_silent(model)
           set_string_names_on_creation(model, false)
           @variable(model, x[i in 1:10_000] in MOI.Interval(0.0, 1.0))
           return model
       end
main (generic function with 1 method)

julia> @time main();
  0.098376 seconds (82.31 k allocations: 7.882 MiB, 63.35% compilation time)

julia> using BenchmarkTools

julia> @benchmark main()
BenchmarkTools.Trial: 156 samples with 1 evaluation.
 Range (min … max):  30.089 ms … 36.249 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     31.996 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   32.061 ms ±  1.038 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

           ▃  ▃▃▅▅ ▂ ▃▆█ ▆ ▆▃                                  
  ▄▄▅▅▄▇▇▅▄█▅███████████▅████▇██▇▇█▅▅▅▄▁▄▁▄▄▁▁▄▄▁▁▁▁▄▁▄▁▁▁▁▁▄ ▄
  30.1 ms         Histogram: frequency by time        35.5 ms <
```
to
```Julia
julia> using JuMP, HiGHS

julia> function main()
           model = direct_model(HiGHS.Optimizer())
           set_silent(model)
           set_string_names_on_creation(model, false)
           @variable(model, x[i in 1:10_000] in MOI.Interval(0.0, 1.0))
           return model
       end
main (generic function with 1 method)

julia> @time main();
  0.087588 seconds (67.48 k allocations: 6.796 MiB, 82.41% compilation time)

julia> using BenchmarkTools

julia> @benchmark main()
BenchmarkTools.Trial: 323 samples with 1 evaluation.
 Range (min … max):  12.555 ms … 297.110 ms  ┊ GC (min … max):  0.00% … 92.31%
 Time  (median):     13.754 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):   15.470 ms ±  16.230 ms  ┊ GC (mean ± σ):  10.89% ± 12.55%

  ▄▆▇█▇▄                                                        
  ███████▇▄▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▅▇▄▁▅▄▁▄▁▁▁▁▁▄ ▆
  12.6 ms       Histogram: log(frequency) by time      33.9 ms <

 Memory estimate: 4.90 MiB, allocs estimate: 40064.
```